requirements.txt
attrs==23.2.0
Automat==22.10.0
click==8.1.7
constantly==23.10.4
# Editable install with no version control (cryptography==0.1.0)
-e /home/arthur/Documentos/Programacao/Python/criptograph
hyperlink==21.0.0
idna==3.7
ifaddr==0.2.0
incremental==24.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
rich==13.7.1
setuptools==71.1.0
six==1.16.0
Twisted==24.3.0
typing_extensions==4.12.2
zeroconf==0.132.2
zope.interface==6.4.post2
